requirements.txt
streamlit
pandas
numpy
scikit-learn
matplotlib
seaborn
pyyaml
kafka-python
joblib
